+++ library/src/methods/strict/index.ts
@@ -1,0 +1,2 @@
+export * from './strict.ts';
+export * from './strictAsync.ts';